requirements.txt
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-mgmt-compute==33.0.0
azure-mgmt-core==1.4.0
azure-mgmt-network==27.0.0
azure-mgmt-resource==23.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
idna==3.10
isodate==0.7.2
lark==1.2.2
msal==1.31.0
msal-extensions==1.2.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-hcl2==5.1.1
pywin32==308
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
